requirements.txt
Django==5.2.5
djangorestframework
requests
gunicorn
